requirements.txt
astor==0.8.0
autograd==1.3
blinker==1.4
boto3==1.14.31
botocore==1.17.31
brotli==1.0.7
bz2file==0.98
ca_certs_locater==1.0
chromedriver_binary==85.0.4183.38.0
click==7.1.2
cloud_tpu_client==0.10
cloudpickle==1.5.0
contextlib2==0.6.0.post1
cryptography==3.0
Cython==0.29.21
dataclasses==0.7
Dialog==0.2.0
diff==0.6.1
disco==1.40.4
Django==3.0.8
docutils==0.16
facets_overview==1.0.0
fasteners==0.15
fastparquet==0.4.1
Flask==1.1.2
freezegun==0.3.15
funcsigs==1.0.2
functools32==3.2.3-2
genapi==0.0.8
google_api_python_client==1.10.0
google_apitools==0.5.31
gprof2dot==2019.11.30
greenlet==0.4.16
grpc_health==9.9.0
grpc_reflection==1.0.0
grpc_tools==1.0.0
guppy==0.1.11
h5==0.0.1
hypothesis==5.23.7
importlib_resources==3.0.0
ipaddr==2.2.0
ipython==7.16.1
ipywidgets==7.5.1
jupyter_client==6.1.6
keyring==21.2.1
kubernetes==11.0.0
lxml==4.5.2
lz4==3.1.0
Mako==1.1.3
matplotlib==3.3.0
memory_profiler==0.57.0
monotime==1.0
mpi4py==3.0.3
mpmath==1.1.0
mtrand==0.1
nbconvert==5.6.1
nbformat==5.0.7
needle==0.5.0
nose==1.3.7
numarray==1.5.1
Numeric==24.2
ordereddict==1.1
pandas==1.1.0
pathlib2==2.3.5
pickle5==0.0.11
Pillow==7.2.0
portpicker==1.3.1
pox==0.2.8
psutil==5.7.2
pycryptodome==3.9.8
pydot_ng==2.0.0
pydotplus==2.0.2
pydub==0.24.1
pyfakefs==4.1.0
Pygments==2.6.1
PyHamcrest==2.0.2
pymongocrypt==1.0.1
pyOpenSSL==19.1.0
pytest==6.0.1
pytidylib==0.3.2
PyYAML==5.3.1
reno==3.1.0
requests_kerberos==0.12.0
requests_mock==1.8.0
scikits.talkbox==0.2.5
secrets==1.0.2
selenium==3.141.0
setuptools_scm==4.1.2
simplejson==3.17.2
Sphinx==3.1.2
spline==1.12
streams==0.1
sympy==1.6.1
tblib==1.7.0
tenacity==6.2.0
tensorboard_plugin_profile==2.3.0
tensorflow_data_validation==0.22.2
tensorflow_model_analysis==0.22.2
tensorflow_transform==0.22.0
testrepository==0.0.20
testresources==2.0.1
testscenarios==0.5.0
testtools==2.4.0
timeloop==1.0.2
unittest2==1.1.0
virtualenv==20.0.28
watchdog==0.10.3
webapp2==2.5.2
wincertstore==0.2
winkerberos==0.7.0
zstandard==0.14.0
